Pulse Ability Implemented
NOTE: Lights must now be added to Players shadow detection script in order to work.

Pulse abilty is now implemented. When pulse button is activated and player can pulse, the Player a pulse corroutine. The corutine will turn on the pulse point light and increase its range over time and then shrink at the same rate. (NOTE: Rate can be altered by changing the pulseRate in Player script. ) Then the coroutine waits out the pulse cooldown time and then it will allow the player to pulse again.

I want to adjust the sudden jerk in player size that occurs right when method is launched due to the shadowLife cost for preforming the pulse. 

Fixed issue with crouch animation not playing.